Fix visual mode movements ignoring relative cursor location
Previously the code always considered the upper visual mode end to be
the origin of the visual mode selection (cursor at bottom) and
completely disregarded that it could be the other way around (cursor at
upper end of the selection).

Theoretically this could also have been fixed by running "gvo" a second
time to get the origin after the first "gv" to get the cursor location.
But that would have implied an unnecessary jump that would have to be
restored (including fixing the window view port), so it's easier to just
get the current cursor location and compare that with the visual mode
end markers to know if the cursor is at the upper or lower end of the
selection.