Hold up invariant of no duplicate hints
Due to a bug in the hint generation it was possible to end up with
duplicate hints.
This only happened when the number of hints exceeded n^2 + 1, where n is
the number of possible hint characters.

The problem was that [:-1] is a special case for slices and reverts back
to the last character of the array so returning the whole array, instead
of returning the empty array as expected.