Distribute keys evenly
The wrap-around was intended from the beginning on, but didn't work
until now, because the modulo of negative numbers is still the same
negative number.

To fix this, always make sure the number is positive before applying
modulo.

The uniform distribution is important when generating a lot of hint
keys.

Before this patch, the first key would always be chosen, which means
that the hint length would explode for one specific prefix.

For example you would get 90% of keys like "cccccccccccccccca" with 10%
of keys being very short like "ab".

With this patch the length is evenly distributed, so you have 100% of
keys like "cab".

This also adds a test to check for prefix collision to make sure our
algorithm still works correctly.