Use a single dict option for all options
This is much more saner than using a separate variable for every option,
particularly because it is more discoverable and easier to document.
If the user sets only some of the options, the extend() makes sure that
the rest is still retrieved from the default options.

Since I am probably the only user of this plugin, this breaking change
should be fine. I would still like to introduce this option system in
all other plugins, but it will be difficult if it breaks the config of
some users.