Add a benchmark for showing hints
Vim at the moment does not offer an extra API for showing virtual text
overlays, so we have to create a popup for every hint [0]. Once the
number of popups becomes somewhat large (in this benchmark 4096), a
performance problem slowly starts appearing.

This patch is meant to act as a first step for profiling how we can
improve this performance problem in Vim.

[0] https://github.com/vim/vim/issues/15007