Avoid visual mode dichotomy
By using the special <Cmd> mapping instead of a manual :call, we can
avoid the unneeded switch to normal mode, in case the function is
invoked from visual mode.

Luckily the rest of the logic does not assume any specific mode, so now
we finally have the same codepath for both visual and normal mode and do
no longer need to pass the visual mode around or even restore it.